yt:git_commit:cb89924b8e1466b70f504a1b1afcfa2407678f7c

Note: forcefully committed, non-critical svn checks were bypassed (txn id 3554898-26vr8)

git-svn-id: svn+ssh://arcadia.yandex.ru/arc/trunk/arcadia/yt/19_3/yt@3554900 41d65440-b5be-11dd-afe3-b2e846d9b4f8

+++ ytlib/misc/memory_chunk_provider-inl.h
@@ -83,9 +83,9 @@ void TPooledObjectTraits<TTrackedAllocationHolder<Size, PoolCategory>>::Clean(TT
 }
 
 template <size_t Size, NNodeTrackerClient::EMemoryCategory PoolCategory>
-size_t TPooledObjectTraits<TTrackedAllocationHolder<Size, PoolCategory>>::GetMaxPoolSize()
+int TPooledObjectTraits<TTrackedAllocationHolder<Size, PoolCategory>>::GetMaxPoolSize()
 {
-    return std::numeric_limits<size_t>::max();
+    return std::numeric_limits<int>::max();
 }
 
 ////////////////////////////////////////////////////////////////////////////////
